requirements.txt
fastapi~=0.103.1
uvicorn~=0.23.2
aiohttp~=3.8.5
pymongo~=4.5.0
pydantic~=2.3.0
pydantic_settings~=2.0.3
motor~=3.3.1
python-dotenv~=1.0.0